Move createMailTo
