+++ packages/react-components/src/index.ts
@@ -5,8 +5,7 @@ import * as authTestUtils from './auth-test-utils';
 
 export { messages, pixels, text, authTestUtils };
 
-export { createMailTo } from './utils';
-
+export * from './mail';
 export * from './icons';
 export * from './images';
 export * from './theme';
