+++ packages/react-components/src/templates/TeamAbout.tsx
@@ -5,7 +5,7 @@ import { TeamResponse } from '@asap-hub/model';
 import { perRem } from '../pixels';
 import { MembersSection, ProfileSkills, TeamOverview } from '../organisms';
 import { CtaCard } from '../molecules';
-import { createMailTo } from '../utils';
+import { createMailTo } from '../mail';
 
 const styles = css({
   display: 'grid',
